Login

+++ src/main/java/com/gameduca/repository/AsignaturaRepository.java
@@ -3,6 +3,6 @@ package com.gameduca.repository;
 import com.gameduca.entity.Asignatura;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AsignaturaRepository extends CrudRepository<Asignatura, Integer>{
+public interface AsignaturaRepository extends CrudRepository<Asignatura, Long>{
 
 }
